(PE-6019) Set request body character encoding to UTF-8
This commit sets the character encoding for the body of a request to
UTF-8.  Previous code did not set a character encoding and so a default
of ISO-8859-1 was implicitly used.  This change assumes that UTF-8 is a
more reasonable default for clients although work should be done later
to make this externally configurable.