(TK-97) Switch from .setSSLContext to .setSSLStrategy
Prior to this commit, we were using `HttpAsyncClientBuilder.setSSLContext`
to set up our HTTPS SSL configuration.  That method doesn't provide
a mechanism for specifying a list of acceptable protocols or ciphers.

This commit simply ports the code over to use `.setSSLStrategy`
instead, because it allows a larger number of configuration options.