(TK-103) More DRYing up and restored some options documentation
This commit removes the `request-with-reified-client` function in
async.clj in lieu of having clients just call the `request-with-client`
function.

This commit also restores documentation for the client and request
options to the `create-client` and `request-with-client` functions,
respectively.  The documentation includes some updates for newer options
which are supported but not were not included in the code currently on
the master branch.