(TK-179) Have JavaClient/createClient do coercion of client options
Previously, anything that called `JavaClient/createClient` performed the same
coercion on the client options before passing them in as the argument.  This
commit refactors `createClient` to do this coercion itself, thereby removing a
bunch of duplication.

Also, remove the `common/Client` schema and replace it with the
`HttpAsyncClient` interface, and remove an unnecessary call to `.start` in the
clojure `create-client`.