(TK-316) Fix metrics for requests with unbuffered streaming
Previously, requests that used unbuffered streaming would get counted twice in
the request metrics. This commit fixes this issue by creating a second metric
for unbuffered streaming requests that times the streaming of the payload
(stopping the timer when the tream returned by the remote server is completely
consumed by Apache async), in addition to the "normal" metric on the request
timing when its response is received. This timer on this "normal" metrics
times is stopped as soon as the resonse header is available and the stream is
ready to consume.