(TK-316) Rename timers and remove 'response-init' timer
Update metrics namespace to include "experimental".

Rename the 'unbuffered_stream' timer to 'bytes-read' and remove the timer
previously refered to as "normal." This timer was really timing when the
response was initially received. In the future, we would like to have a timer
timing this, however at the moment we can only easily get this for the
unbuffered stream requests. We will reintroduce this when we can include this
timer on all requests.