(TK-316) Make metric registry a client option
Previously, when creating a client (with both Java and Clojure), a metric
registry was passed in as an optional second argument to the
`createClient()`/`create-client` functions. However, everything else that was
used as an option was passed in in the options map (Clojure) or ClientOptions
(Java). This commit makes the metric registry option the same as all the other
options, rather than a second argument to the function.