(TK-316) Distinguish between metric id and metric name
Previously, the term 'metric id' was used to mean both an array request option
used to create a metric name to time the request and the metric name itself.
This commit separates these uses out, so that 'metric id' is used to refer to
this array of options, and 'metric name' refers to the timer name.