(TK-316) Add with-url-and-method namespace
Move the timers that have both the url and the method from the `with-url`
namespace to a new `with-url-and-method` namepsace. This will better match up
with the data structure they will later be returned in and the filtering
mechanisms for getting them back out, and also removes a slight chance of
having a conflict with a url timer.