(TK-316) Add ClientTimer and getClientMetricsBy* methods
Add a ClientTimer class that wraps the Timer class and also holds onto the
url, method, or metric id used to create the timer.

Remove the ClientMetricRegistry which was previously doing some of this, but
is now no longer needed.

Use the ClientTimer to add url, method, metricId to ClientMetricData.

Change the getClientMetrics* methods to return a map of metric category to
array of timers or timer data. Add getClientMetrics(Data)By* methods that
allow for filtering by a specific metric category and return arrays of timers
or timer data.

Remove the filter-builder functions as they are no longer needed.