(PDB-2640) Close gzip-wrapped body input stream on request completion
In the previous commit, if gzip compression was applied to a request
input stream, the stream was not automatically closed when the request
was complete.  The Apache HTTP async library, however, closes an input
stream in cases where gzip compression is not being added.  In this
commit, the original request input stream is closed immediately after
attempts to copy the data from it to the downstream gzipped input stream
have been completed, for compatibility with the pre-existing Apache HTTP
async library behavior.