(maint) Update testing pki
Our tests use a CA and node cert, plus an untrusted CA cert. All of
these had different expirations and one had recently expired.

This commit updates all of them to have an expiration in 5 years as well
as creates a script to ease managing them and updates the README to
point towards users towards the script.