(maint) update clj-parent and fix new failures
THis updates clj-parent to 1.7.12, which brings in new versions of
the apache libraries.  There are some functional diffeerences in
the behavior of the library, which have manifested themselves in
places where this library is used.  The tests were updated to
reflect the current behavior, and the JavaClient was updated to
avoid throwing an exception when the content-type is malformed.

Some formatting differences in the content-type header responses
were also changed.

The travis file was updated to reflect the correct version of
java.