(PE-25106) add support for max-per-route and max-total
This adds support for configuration of a couple of parameters for
perisistent clients.  Internally, the apache http client uses
a per-route (host + port) limit of 2 concurrent connections, and
a maximum total of 20 concurrent connections.

This adds configuration values to allow both to be set, and
tests that demonstrate how they behave.

In addition, the docs were updated, and the changelog updated to
reflect the expected new release number, as this is a feature addition.