(PDB-4357) FIPS support via bc-fips jar
Move the bouncycastle jar to the dev profile. This will allow us
to specify the bouncycastle jars on the classpath at runtime like
we do for logging configuration. At installation time, we can install
the FIPS-compliant version of the jars, or the standard versions.

Remove any and all SSLv3 references / tests. FIPS does not allow SSLv3,
and the JVM explicitly disables SSLv3. OWASP guidelines specifically
state:

    SSL/TLS is a collection of protocols. Weaknesses have been identified with earlier SSL protocols, including SSLv2 and SSLv3, hence SSL versions 1, 2, and 3 should not longer be used. The best practice for transport layer protection is to only provide support for the TLS protocols - TLS 1.0, TLS 1.1 and TLS 1.2

Removing the test.sh script in favor of calling lein directly.

Adding CI tests for JDK 8, with and without FIPS.

Removing the need for nil version specifiers by bumping the minimum lein
version to 2.9.1, and updating lein-parent to 0.3.7

bump clj-parent to 4.0.0 to pick up jvm-ssl-utils