(PE-27189) disambiguate ambigous types
There have been issues seen with clj and java11 where clojure is unable
to find the class associated with specific java interop calls.  In this
case, the "RequestOptions" constructor was occasionally ambigous.  In
order to rectify this issue, in all places where there is some ambiguity,
type assertions are added.

In addition, there was some type issues with the common/RequestOptions
ahd how it was handled.  With the exiting code, there is one consistent
path through RequestOptions that passes a URI.  Schema type assertions
were added to demonstrate and validate the behavior, and tests were
created to demonstrate the correct handling of URL and URI types for
url specification.