(SERVER-3068) Filter auth headers on redirect
When being redirected all headers are copied from the original request
to the redirected request. There are some concerns about the security
impact of copying authorization headers to a redirected location.

This implements subclasses of RedirectStrategies that we use internally,
SafeDefaultRedirectStrategy and SafeLaxRedirectStrategy respectively
extend DefaultRedirectStrategy and LaxRedirectStrategy. The redirect
spec is technically only valid for HEAD and GET verbs and is the
behavior of the DefaultRedirectStrategy, while the LaxRedirectStrategy
allows redirecting for any verb, except for PUT, and is used if
force-redirects is enabled.

The subclassed RedirectStrategies prevent non-GET/HEAD verbs from having
their authorization headers copied and wrap all redirected requests in a
SafeRedirectedRequest. A SafeRedirectedRequest proxies all method
invocations to the wrapped request except for setHeaders(), which is
used by callers to copy over headers from the original request (yes, in
the case of non-GET/HEAD verbs, the headers are copied twice, with the
second copy overwriting the first).

These code paths are deprecated, do not exist in 5.0 of the library and
change in important ways between patch versions within the 4.5 series.
It is important to audit these classes whenever upgrading dependencies.