(SERVER-3068) Conditionally filter auth headers
Previously, we unconditionally filtered out `Authorization` and
`X-Authorization` headers when following redirects. This approach caused
issues because internal services rely on that behavior.

A better solution, what we believe browsers implement, is to allow auth
headers if the request matches the same scheme, port, and host. The
acceptance criteria for the ticket was updated to allow redirects to
include the auth headers for any subdomain under the same domain that
initiated the redirect as well as allow upgrading from http to https. In
addition the scope of "auth headers" was extended to include `Cookie`,
`Set-Cookie`, `WWW-Authenticate`, `Proxy-Authenticate`, and
`Proxy-Authorization`.

This commit extends the scope of auth headers and allows them to be
propagated to redirects only if the scheme, host, and port match between
the original target and the redirect. The more restricted criteria is in
the interest of expediency since we do not use subdomains ourselves.