(PE-33177) Update Travis to use more recent builds for java 8 and 11
The versions of Java Travis has pre-installed are pretty old. The sync_ssl_test is looking for particular exceptions that have changed since these old versions. This changes the instructions for Travis to install OpenJDK from the Ubuntu repos, which are more up to date (though still not latest).

This also modifies the jobs to run in parallel, rather than sequentially, since they don't depend on each other. This wil save some time.