(maint) Speed up unit tests
Some of the tests in async-plaintext-test were leaving the server
waiting for a countdown latch that was never completed, causing the
shutdown to timeout after 30 seconds.

The blocking tests in async-unbuffered-test had a similar issue where
the server was hitting a 1 second timeout before shutting down. They
were also using far larger test data (32MB) than necessary. Since that
test data was populated via UUID generation, it was quite expensive both
to generate it in addition to the cost of actually transferring it over
HTTP. We now deliver a promise at the end of the test indicating the
blocking handler should complete, and we only use 1MB of data which is
plenty to saturate the client's buffer.