(maint) add support for multiple Set-Cookie headers
The design of the client does not allow multiple entries for the
same header name, as headers are returned in a map, rather than
an array of them.  As a result, headers like "Set-Cookie" would only
return the last one seen if they were encountered. This alters the
behavior to concatenate any of the "Set-Cookie" entries together,
separated by a newline character. This seems to be a fairly common
convention for clients that use headers with maps.

Tests were updated to include multiple cookies, and demonstrate
the behavior.