(maint) Move failure logging to debug level
Some services in PE poll services frequently and expect a high number of
failures to occur during normal operations. This causes costly and
concernly levels of error messages.

This moves the default for the failure messages added in 898ccc01f to the
debug level. Users are advised to raise the log level of
`com.puppetlabs.http.client.impl.JavaClient` to debug to see the output
of spurious failures.